Rename component from LogInForm to LoginForm
