--- src/components/LogInForm/index.tsx
@@ -1,1 +1,0 @@
-export { LogInFormContainer as LogInForm } from "./container";
