+++ src/components/LoginForm/index.tsx
@@ -1,0 +1,1 @@
+export { LoginFormContainer as LoginForm } from "./container";
